Apply comments. Move plugin cache to common test utils
