+++ src/plugins/intel_cpu/tests/functional/custom/subgraph_tests/src/input_tensor_roi.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "functional_test_utils/ov_plugin_cache.hpp"
+#include "common_test_utils/ov_plugin_cache.hpp"
 #include "utils/cpu_test_utils.hpp"
 
 using namespace CPUTestUtils;
