+++ src/plugins/intel_gpu/tests/functional/behavior/memory_dyn_batch.cpp
@@ -6,7 +6,7 @@
 #include "common_test_utils/common_utils.hpp"
 #include "common_test_utils/test_constants.hpp"
 #include "functional_test_utils/skip_tests_config.hpp"
-#include "functional_test_utils/ov_plugin_cache.hpp"
+#include "common_test_utils/ov_plugin_cache.hpp"
 #include "openvino/core/partial_shape.hpp"
 #include "openvino/runtime/compiled_model.hpp"
 #include "openvino/runtime/infer_request.hpp"
