+++ src/tests/functional/plugin/shared/include/behavior/ov_infer_request/infer_request_dynamic.hpp
@@ -11,7 +11,7 @@
 #include <vector>
 #include <string>
 #include <memory>
-#include "functional_test_utils/ov_plugin_cache.hpp"
+#include "common_test_utils/ov_plugin_cache.hpp"
 #include "openvino/core/shape.hpp"
 #include "shared_test_classes/base/layer_test_utils.hpp"
 #include <string>
