+++ src/tests/functional/plugin/shared/src/behavior/ov_infer_request/inference.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 #include "openvino/opsets/opset8.hpp"
-#include "functional_test_utils/ov_plugin_cache.hpp"
+#include "common_test_utils/ov_plugin_cache.hpp"
 #include "behavior/ov_infer_request/inference.hpp"
 
 namespace ov {
