+++ src/tests/functional/plugin/shared/src/execution_graph_tests/add_output.cpp
@@ -4,7 +4,7 @@
 
 #include "common_test_utils/common_utils.hpp"
 
-#include "functional_test_utils/ov_plugin_cache.hpp"
+#include "common_test_utils/ov_plugin_cache.hpp"
 #include "functional_test_utils/skip_tests_config.hpp"
 
 #include "execution_graph_tests/add_output.hpp"
