+++ src/tests/test_utils/common_test_utils/src/ov_plugin_cache.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "functional_test_utils/ov_plugin_cache.hpp"
+#include "common_test_utils/ov_plugin_cache.hpp"
 
 #include <gtest/gtest.h>
