+++ src/tests/test_utils/common_test_utils/src/ov_test_utils.cpp
@@ -5,6 +5,7 @@
 #include "common_test_utils/ov_test_utils.hpp"
 
 #include "common_test_utils/file_utils.hpp"
+#include "common_test_utils/ov_plugin_cache.hpp"
 #include "common_test_utils/test_constants.hpp"
 #include "openvino/runtime/core.hpp"
 #include "openvino/util/file_util.hpp"
@@ -119,32 +120,23 @@ namespace ov {
 namespace test {
 namespace utils {
 
-ov::TensorVector infer_model_via_template(const std::shared_ptr<ov::Model>& model,
-                                          const ov::TensorVector& input_tensors) {
+ov::TensorVector infer_on_template(const std::shared_ptr<ov::Model>& model, const ov::TensorVector& input_tensors) {
     std::map<std::shared_ptr<ov::Node>, ov::Tensor> inputs;
     auto params = model->inputs();
     OPENVINO_ASSERT(params.size() == input_tensors.size());
     for (int i = 0; i < params.size(); i++) {
         inputs[params[i].get_node_shared_ptr()] = input_tensors[i];
     }
-    return infer_model_via_template(model, inputs);
+    return infer_on_template(model, inputs);
 }
 
-ov::TensorVector infer_model_via_template(const std::shared_ptr<ov::Model>& model,
-                                          const std::map<std::shared_ptr<ov::Node>, ov::Tensor>& inputs) {
-    ov::Core core;
+ov::TensorVector infer_on_template(const std::shared_ptr<ov::Model>& model,
+                                   const std::map<std::shared_ptr<ov::Node>, ov::Tensor>& inputs) {
+    auto core = ov::test::utils::PluginCache::get().core();
 
-    // Register Template plugin
-#ifndef OPENVINO_STATIC_LIBRARY
-    std::string pluginName = ov::test::utils::TEMPLATE_LIB;
-    pluginName += OV_BUILD_POSTFIX;
-    core.register_plugin(ov::util::make_plugin_library_name(ov::test::utils::getExecutableDirectory(), pluginName),
-                         ov::test::utils::DEVICE_TEMPLATE);
-#endif  // !OPENVINO_STATIC_LIBRARY
-
-    auto compiled_model = core.compile_model(model,
-                                             ov::test::utils::DEVICE_TEMPLATE,
-                                             {{ov::template_plugin::disable_transformations(true)}});
+    auto compiled_model = core->compile_model(model,
+                                              ov::test::utils::DEVICE_TEMPLATE,
+                                              {{ov::template_plugin::disable_transformations(true)}});
     auto infer_request = compiled_model.create_infer_request();
 
     for (auto& input : inputs) {
